pyproject.toml
[tool.poetry]
name = "slotBooker"
version = "2.7.1"
description = ""
authors = ["Sebastian Blum <hello@seblum.com>"]
packages = [{include = "slotbooker", from = "src"}]


[tool.poetry.plugins.dotenv]
ignore = "false"
location = ".env"

[tool.poetry.dependencies]
python = "^3.11"
selenium = "^4.6.0"
DateTime = "^4.5"
PyYAML = "^6.0"
tqdm = "^4.64.0"
termcolor = "^2.3.0"
poetry-plugin-dotenv = "^2.1.6"
coverage = "^7.5.3"
retrying = "^1.3.4"
tenacity = "^9.1.2"

[tool.poetry.group.dev.dependencies]
pytest = "^8.2.2"
ruff = "^0.7.1"
coverage = "^7.5.3"

[tool.poetry.scripts]
slotBooker = "slotbooker.main:main"
slotBookerDev = "slotbooker.main:development"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.ruff]
lint.select = ["E", "F", "W", "C", "N", "Q"]
lint.ignore = ["E501"]  # ignore line length for example
line-length = 88

# Exclude some directories from linting
exclude = ["migrations", "node_modules", "venv"]
# Configure fix options
# fixable = true
# # Configure per-file-ignores
# per-file-ignores = [
#     "tests/test_*.py: F401, F403"
# ]
